Add script to migrate asset keys.
We want to convert data in our level DB database to a new format.
Our original format made it very difficult to view all the assets in a
given bucket or to view the most current asset in a given bucket in any
programatic way. After running this migration keys in the form:
`bucket!![bucketId]!![assetName]!![assetId]`
will be in the form:
`bucket-asset-version!![bucketId]!![assetName]!![assetId]`

and keys in the form:
`bucket!![bucketId]!!asset!![assetName]`
will be in the form:
`bucket-asset!![bucketId]!![assetName]`

Additionaly the filename of the asset is added to the asset value.
Ex:
asset!!54ac80317d800000 -> {
  "token": "78edd575-39fd-44c9-bf1b-b09ec5d261ed",
  "time":1454685341175,
  "fileName":"eight.sql"
}