Working script to recipher files & get file sizes.
We want to track the gzipped and raw sizes of all files uploaded to the
asset receiver. This is in place however we need to capture this data
for older files. This migration script captures both pieces of data for
all files managed by the asset receiver and updates the metatdata in
levelDB. At the same time we also recipher all the files to be salted
thus adding a security measure.