Change server cipher password.
From here out we are going to use a password to encrypt file assets that
relies on attributes of the file itself and the bucket mixed with our
own psuedo-random base. This will make it more difficult for an attacker
to decrypt client data if he is able to escalate. The migration already
handles re-ciphering old data and this change ensures we use the same
system going forward.