Fix asset size test & create asset size endpoint.
The asset size test was failing because it relied on the old format of
levelDB keys. In order to actually test the code properly, we need a new
endpoint that will return the asset metadata based on the asset name and
the bucket. This adds the necessary endpoint and adjusts the test
accordingly.