API endpoint to list metadata of bucket assets.
We need an API endpoint that will return all the metadata for all the
assets associated with a bucket so that users can decide which ones to
keep and which ones to remove since we will be limiting the amount of
disk space they can use. This provides the needed endpoint and tests for
it.