Add migration to remove problematic keys.
In some cases we may have keys that could not be updated by previous
migrations. Anything in the form `bucket!![bucketId]!![ANYTHING]
will cause problems when we are trying to list buckets from the API. To
resolve this, after running the other migrations, this script will
remove all of the problematic keys.