Fix AWS asset receiver upload connection end.
The asset upload endpoint would not properly fire an end event when
working with the AWS S3 plugin. The cause of this was that we were
piping the file data through to the S3 uploader but our return string
had no way to detect that the S3 API's done callback had finished. Here
we fix the issue by closing the stream when S3 reports that the file has
successfully uploaded.