Delete asset versions when deleting asset.
We had been leaving historical asset versions on disk and in the LevelDB
database after deleting an asset. This was not the desired functionality
as it would eventually result in us having to run migrations to remove
old asset versions. Currently we do not give users a way to manage these
so it can be assumed that a user will expect all data related to an
asset to be removed upon deletion and this includes historical versions.

While testing, I found an issue in which the AWS uploader could fail and
incorrectly report to the uploader that it had succeeded. To deal with
this condition I added an error handler on the stream and ensured the
the AWS upload function emits an error event when relevant.