Allow for asset uploads to be suspended.
In some cases we need to be able to stop new assets from being
uploaded but would like to allow assets to continue to be
downloaded. This is useful for doing server maintenance. This
allows a flag to be set in the config or set via a post to a
special URL. When the "paused" flag is set, the server will
return a 503 resource unavailable response for POSTs to the asset
upload endpoint. Tests are provided for all cases.